Fix TERM=screen-256 flight info rendering issue
The 'srg0' setting does not work with `less -FRX` for all settings.

This is particularly important for screen-256 color which uses
"shift-in" to control font changes. This however confuses less -FRX
causing the rendering issue.

Instead the default SRG can be manually triggered with: \E[;10m

NOTE: The original handling is still required for "dum" (smarter?)
pagers such as 'more'.