Check locale for interactive Flight Environment
Previously, we checked the locale as part of the standard profile
scripts and we did so for both interactive and non-interactive shells.

Now we check the locale only when the flight environment is started and
only if the shell is interactive.

This is in better keeping with Flight being intended to keep out of the
way of the user unless they choose to activate it.  It also prevents
double reporting on bad locales that can happen under some
circumstances.